Rename the package and associated docs, tests
