+++ .gitignore
@@ -107,3 +107,6 @@ ENV/
 
 # Mac
 *.DS_Store
+
+biolm.egg-info/
+docs/_build/
