+++ README.md
@@ -19,7 +19,7 @@ Finally, run the Makefile to complete the setup:
 make install
 ```
 
-This installs the biolmai package itself in editable mode. This allows you to make changes to the package and see those changes reflected without needing to reinstall the package:
+This installs the biolm package itself in editable mode. This allows you to make changes to the package and see those changes reflected without needing to reinstall the package:
 It runs the command: `pip install -e .`. The -e flag stands for "editable mode". This is a useful setup for development as it means any changes you make to the package code will be immediately available in your environment, without needing to reinstall the package each time you make a change.
 
 
@@ -29,7 +29,7 @@ Set the token in your shell to authenticate your requests to the BioLM API:
 export BIOLMAI_TOKEN=<your_token>
 ```
 The token can be obtained from your BioLM account: [Get your BioLM API token](https://biolm.ai/ui/accounts/user-api-tokens/).
-Run `biolmai status` to verify that the token is set correctly.
+Run `biolm status` to verify that the token is set correctly.
 
 ## Run tests with a specific seed
 Use this command to run the automated tests:
