+++ README.rst
@@ -3,8 +3,11 @@ BioLM AI
 ========
 
 
-.. image:: https://img.shields.io/pypi/v/biolmai.svg
-        :target: https://pypi.python.org/pypi/biolmai
+.. image:: https://img.shields.io/pypi/v/biolm.svg
+        :target: https://pypi.python.org/pypi/biolm
+
+.. image:: https://img.shields.io/pypi/v/biolm.svg
+        :target: https://pypi.python.org/pypi/biolm
 
 .. image:: https://api.travis-ci.com/BioLM/py-biolm.svg?branch=production
         :target: https://travis-ci.org/github/BioLM/py-biolm
@@ -16,22 +19,31 @@ BioLM AI
 
 
 
-Python client and SDK for https://biolm.ai
+Python client and SDK for `BioLM <https://biolm.ai>`_
 
+.. include:: docs/python-client/quickstart.rst
+   :start-line: 1
+   :end-line: 27
+
+Asynchronous usage:
+
+.. code-block:: python
+
+    from biolm.client import BioLMApiClient
+    import asyncio
+
+    async def main():
+        model = BioLMApiClient("esmfold")
+        result = await model.predict(items=[{"sequence": "MDNELE"}])
+        print(result)
+
+    asyncio.run(main())
+
+.. include:: docs/python-client/overview.rst
+   :start-line: 1
+   :end-line: 17
+
+.. include:: docs/python-client/features.rst
 
 * Free software: Apache Software License 2.0
 * Documentation: https://biolm-ai.readthedocs.io.
-
-
-Features
---------
-
-* TODO
-
-Credits
--------
-
-This package was created with Cookiecutter_ and the `audreyr/cookiecutter-pypackage`_ project template.
-
-.. _Cookiecutter: https://github.com/audreyr/cookiecutter
-.. _`audreyr/cookiecutter-pypackage`: https://github.com/audreyr/cookiecutter-pypackage
