+++ biolm/__init__.py
@@ -3,8 +3,8 @@ __author__ = """Nikhil Haas"""
 __email__ = "nikhil@biolm.ai"
 __version__ = '0.2.3'
 
-from biolmai.cls import *
-from biolmai.client import BioLMApi, BioLMApiClient
-from biolmai.biolmai import BioLM
+from biolm.cls import *
+from biolm.client import BioLMApi, BioLMApiClient
+from biolm.biolmai import BioLM
 
 __all__ = []
