+++ biolm/api.py
@@ -14,12 +14,12 @@ import requests
 from requests.adapters import HTTPAdapter
 from requests.packages.urllib3.util.retry import Retry
 
-import biolmai
-import biolmai.auth
-from biolmai.asynch import async_api_call_wrapper
-from biolmai.biolmai import log
-from biolmai.const import MULTIPROCESS_THREADS
-from biolmai.payloads import INST_DAT_TXT, PARAMS_ITEMS, predict_resp_many_in_one_to_many_singles
+import biolm
+import biolm.auth
+from biolm.asynch import async_api_call_wrapper
+from biolm.biolmai import log
+from biolm.const import MULTIPROCESS_THREADS
+from biolm.payloads import INST_DAT_TXT, PARAMS_ITEMS, predict_resp_many_in_one_to_many_singles
 
 
 @lru_cache(maxsize=64)
@@ -69,7 +69,7 @@ def validate_action(action):
                 name = f"{f.__module__}.{f.__name__}"
 
             if is_method:
-                # Splits name, e.g. 'biolmai.api.ESMFoldSingleChain.predict'
+                # Splits name, e.g. 'biolm.api.ESMFoldSingleChain.predict'
                 action_method_name = name.split(".")[-1]
                 validate_endpoint_action(
                     class_obj_self.action_class_strings,
@@ -166,7 +166,7 @@ class APIEndpoint:
         else:
             self.multiprocess_threads = MULTIPROCESS_THREADS  # Could be False
         # Get correct auth-like headers
-        self.auth_headers = biolmai.auth.get_user_auth_header()
+        self.auth_headers = biolm.auth.get_user_auth_header()
         self.action_class_strings = tuple(
             [c.__name__.replace("Action", "").lower() for c in self.action_classes]
         )
