+++ biolm/cli.py
@@ -1,21 +1,21 @@
-"""Console script for biolmai."""
+"""Console script for biolm."""
 import os
 import sys
 
 import click
 
-from biolmai.auth import (
+from biolm.auth import (
     generate_access_token,
     get_auth_status,
     save_access_refresh_token,
 )
-from biolmai.const import ACCESS_TOK_PATH, BASE_API_URL, MULTIPROCESS_THREADS
+from biolm.const import ACCESS_TOK_PATH, BASE_API_URL, MULTIPROCESS_THREADS
 
 
 @click.command()
 def main(args=None):
-    """Console script for biolmai."""
-    click.echo("Replace this message by putting your code into " "biolmai.cli.main")
+    """Console script for biolm."""
+    click.echo("Replace this message by putting your code into " "biolm.cli.main")
     click.echo("See click documentation at https://click.palletsprojects.com/")
     return 0
