+++ docs/Makefile
@@ -4,7 +4,7 @@
 # You can set these variables from the command line.
 SPHINXOPTS    =
 SPHINXBUILD   = python -msphinx
-SPHINXPROJ    = biolmai
+SPHINXPROJ    = biolm
 SOURCEDIR     = .
 BUILDDIR      = _build
