+++ docs/biolm.rst
@@ -1,0 +1,101 @@
+biolm package
+=============
+
+Submodules
+----------
+
+biolm.api module
+----------------
+
+.. automodule:: biolm.api
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+biolm.asynch module
+-------------------
+
+.. automodule:: biolm.asynch
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+biolm.auth module
+-----------------
+
+.. automodule:: biolm.auth
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+biolm.biolm module
+--------------------
+
+.. automodule:: biolm.biolm
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+biolm.cli module
+----------------
+
+.. automodule:: biolm.cli
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+biolm.client module
+-------------------
+
+.. automodule:: biolm.client
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+biolm.cls module
+----------------
+
+.. automodule:: biolm.cls
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+biolm.const module
+------------------
+
+.. automodule:: biolm.const
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+biolm.ltc module
+----------------
+
+.. automodule:: biolm.ltc
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+biolm.payloads module
+---------------------
+
+.. automodule:: biolm.payloads
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+biolm.validate module
+---------------------
+
+.. automodule:: biolm.validate
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: biolm
+   :members:
+   :undoc-members:
+   :show-inheritance:
