--- docs/biolmai.rst
@@ -1,101 +1,0 @@
-biolmai package
-===============
-
-Submodules
-----------
-
-biolmai.api module
-------------------
-
-.. automodule:: biolmai.api
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-biolmai.asynch module
----------------------
-
-.. automodule:: biolmai.asynch
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-biolmai.auth module
--------------------
-
-.. automodule:: biolmai.auth
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-biolmai.biolmai module
-----------------------
-
-.. automodule:: biolmai.biolmai
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-biolmai.cli module
-------------------
-
-.. automodule:: biolmai.cli
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-biolmai.client module
----------------------
-
-.. automodule:: biolmai.client
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-biolmai.cls module
-------------------
-
-.. automodule:: biolmai.cls
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-biolmai.const module
---------------------
-
-.. automodule:: biolmai.const
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-biolmai.ltc module
-------------------
-
-.. automodule:: biolmai.ltc
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-biolmai.payloads module
------------------------
-
-.. automodule:: biolmai.payloads
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-biolmai.validate module
------------------------
-
-.. automodule:: biolmai.validate
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: biolmai
-   :members:
-   :undoc-members:
-   :show-inheritance:
