+++ docs/conf.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 #
-# biolmai documentation build configuration file, created by
+# biolm documentation build configuration file, created by
 # sphinx-quickstart on Fri Jun  9 13:47:02 2017.
 #
 # This file is execfile()d with the current directory set to its
@@ -111,7 +111,7 @@ html_static_path = ["_static"]
 # -- Options for HTMLHelp output ---------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = "biolmaidoc"
+htmlhelp_basename = "biolmdoc"
 
 
 # -- Options for LaTeX output ------------------------------------------
@@ -135,7 +135,7 @@ latex_elements = {
 # (source start file, target name, title, author, documentclass
 # [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, "biolmai.tex", "BioLM AI Documentation", "Nikhil Haas", "manual"),
+    (master_doc, "biolm.tex", "BioLM AI Documentation", "Nikhil Haas", "manual"),
 ]
 
 
@@ -143,7 +143,7 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [(master_doc, "biolmai", "BioLM AI Documentation", [author], 1)]
+man_pages = [(master_doc, "biolm", "BioLM AI Documentation", [author], 1)]
 
 
 # -- Options for Texinfo output ----------------------------------------
@@ -154,10 +154,10 @@ man_pages = [(master_doc, "biolmai", "BioLM AI Documentation", [author], 1)]
 texinfo_documents = [
     (
         master_doc,
-        "biolmai",
+        "biolm",
         "BioLM AI Documentation",
         author,
-        "biolmai",
+        "biolm",
         "One line description of project.",
         "Miscellaneous",
     ),
