+++ docs/python-client/api_client.rst
@@ -2,12 +2,12 @@
 BioLMApi and BioLMApiClient
 ========================
 
-.. autoclass:: biolmai.client.BioLMApi
+.. autoclass:: biolm.client.BioLMApi
    :members:
    :undoc-members:
    :show-inheritance:
 
-.. autoclass:: biolmai.client.BioLMApiClient
+.. autoclass:: biolm.client.BioLMApiClient
    :members:
    :undoc-members:
    :show-inheritance:
@@ -31,7 +31,7 @@ Direct, flexible access to the BioLM API, supporting both synchronous (`BioLMApi
 
 .. code-block:: python
 
-    from biolmai.client import BioLMApi
+    from biolm.client import BioLMApi
 
     # ESM2-8M: encode a batch
     model = BioLMApi("esm2-8m")
@@ -56,7 +56,7 @@ Direct, flexible access to the BioLM API, supporting both synchronous (`BioLMApi
     result = model._batch_call_autoschema_or_manual("encode", batches)
 
     # Async usage
-    from biolmai.client import BioLMApiClient
+    from biolm.client import BioLMApiClient
     import asyncio
 
     async def main():
