+++ docs/python-client/disk_output.rst
@@ -26,7 +26,7 @@ When you set `output='disk'` and provide a `file_path`, results are written as J
     BioLM(entity="esmfold", action="predict", type="sequence", items=["SEQ1", "BADSEQ"], output='disk', file_path="results.jsonl", stop_on_error=True)
 
     # Advanced: retry failed batches as single items (BioLMApi only)
-    from biolmai.client import BioLMApi
+    from biolm.client import BioLMApi
     model = BioLMApi("esm2-8m", retry_error_batches=True)
     model.encode(items=[{"sequence": "SEQ1"}, {"sequence": "BADSEQ"}], output='disk', file_path="out.jsonl")
