+++ docs/python-client/installation.rst
@@ -12,7 +12,7 @@ To install BioLM AI, run this command in your terminal:
 
 .. code-block:: console
 
-    $ pip install biolmai
+    $ pip install biolm
 
 This is the preferred method to install BioLM AI, as it will always install the most recent stable release.
