+++ setup.py
@@ -47,16 +47,16 @@ setup(
     description="BioLM Python client",
     entry_points={
         "console_scripts": [
-            "biolmai=biolmai.cli:cli",
+            "biolm=biolm.cli:cli",
         ],
     },
     install_requires=requirements,
     license="Apache Software License 2.0",
     long_description=readme + "\n\n" + history,
     include_package_data=True,
-    keywords="biolmai",
-    name="biolmai",
-    packages=find_packages(include=["biolmai", "biolmai.*"]),
+    keywords="biolm",
+    name="biolm",
+    packages=find_packages(include=["biolm", "biolm.*"]),
     test_suite="tests",
     tests_require=test_requirements,
     url="https://github.com/BioLM/py-biolm",
