+++ tests/test_aclient.py
@@ -1,12 +1,12 @@
 import json
 import logging
 
-from biolmai.client import BioLMApi
+from biolm.client import BioLMApi
 
 LOGGER = logging.getLogger(__name__)
 
 import pytest
-from biolmai.client import BioLMApiClient
+from biolm.client import BioLMApiClient
 
 @pytest.fixture(scope='function')
 def model():
