+++ tests/test_schemas.py
@@ -1,6 +1,6 @@
 import pytest
 
-from biolmai.client import BioLMApiClient
+from biolm.client import BioLMApiClient
 
 
 @pytest.mark.asyncio
